fix: rename files of academic management module
